preparations for variant entities
- use floats for damage, not double
- change variant to actual class
- use ForgeRegistries instead of BuiltInRegistries
- some helper methods

+++ src/main/java/team/cqr/cqrepoured/entity/profile/variant/DamageEntry.java
@@ -13,14 +13,14 @@ public record DamageEntry(
 		Optional<DamageCap> damageCap,
 		double minDamage,
 		boolean fireImmune,
-		Map<ResourceLocation, Double> damageTypeMultipliers
+		Map<ResourceLocation, Float> damageTypeMultipliers
 ) {
 	public static final Codec<DamageEntry> CODEC = RecordCodecBuilder.create(instance -> {
 		return instance.group(
 				DamageCap.CODEC.optionalFieldOf("damage-cap").forGetter(DamageEntry::damageCap),
 				Codec.DOUBLE.fieldOf("min-damage").forGetter(DamageEntry::minDamage),
 				Codec.BOOL.fieldOf("fire-immune").forGetter(DamageEntry::fireImmune),
-				Codec.unboundedMap(ResourceLocation.CODEC, Codec.DOUBLE).optionalFieldOf("multipliers", Collections.emptyMap()).forGetter(DamageEntry::damageTypeMultipliers)
+				Codec.unboundedMap(ResourceLocation.CODEC, Codec.FLOAT).optionalFieldOf("multipliers", Collections.emptyMap()).forGetter(DamageEntry::damageTypeMultipliers)
 			).apply(instance, DamageEntry::new);
 	});
 }
